pyproject.toml
[tool.poetry]
name = "crajyminder"
version = "0.1.0"
description = ""
authors = ["cs-gang"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
sanic = "^21.3.2"
python-dotenv = "^0.17.0"
Jinja2 = "^2.11.3"
sanic-session = "^0.7.3"
Async-OAuthlib = "^0.0.9"

[tool.poetry.dev-dependencies]
pre-commit = "^2.11.1"
black = "^20.8b1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
